// src/lib/tuples/actor/resolver.d.ts
/*
  Copyright (C) 2018  Miniverse authors

  This program is free software: you can redistribute it and/or modify
  it under the terms of the GNU Affero General Public License as published by
  the Free Software Foundation, version 3 of the License.

  This program is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  GNU Affero General Public License for more details.

  You should have received a copy of the GNU Affero General Public License
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
*/

import Repository from '../../repository';
import Actor from '../actor';

export function lookup(this: void, repository: Repository, resource: string):
  Promise<any>;

export default class Resolver {
  static fromUsernameAndNormalizedHost(
    repository: Repository, username: string, normalizedHost: string):
      Promise<Actor | null>;
  static fromKeyUri(repository: Repository, keyUri: string):
    Promise<Actor | null>;
}